Rework Explanation.
Initially, each page consisted of different lists of objects such as the prompt, a display bar, and a list of options. For the printing of each page, there was a standard order in which the prompt, display bar and options would be printed. Although this created a uniform layout between pages, this was too constrictive for future pages where prompts and options objects might inter-mix.

To work around this issue, I decided to rework the data members within the page class. The new method of printing a page would simply be iterating through a list of mixed elements to print. option, lineStr (a new object), and the displayBar objects now all extend to a parent class called visualElement. Each page now has a vector of type visualElement storing these varied elements.
Note that though there is one single vector of elements that draws the page via a virtual type printElement() function, we still require a options list and a displayBars list for those objects to have full functionality. (You cannot use the members from within the options or displayBars by accessing through the vector<visualElement> elements, as the cast creates limiting issues).
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Explanation aside the end goal with this movement is to transition our page constructor to just the name and list of elements. page(name, elements).
From the previous working state the constructor consisted of page(name, prompt, options, displayBars).
The current commit has a constructor page(name, elements, displayBars). Meaning prompts and options can now be simply declared within the list of elements.
The next movement is to remove the need for a displayBars list, as the displayBars are already declared in the list of elements. This will be done by comparing the names of the elements in the list with the names of a global list of display bars to .push_back onto the page classes' personal displayBar list.