Prototype for first sideBar
Using escape sequences, we are able to move the cursor of where we print lines. With this, I've added a new sideBar element to the page class. The first prototype of this sideBar has been implemented on the play page. The sidebar displays players from the player list as well as their points.

https://www.lihaoyi.com/post/BuildyourownCommandLinewithANSIescapecodes.html
http://rrbrandt.dee.ufcg.edu.br/en/docs/ansi/cursor